+++ react/o11y/instrument.ts
@@ -1,5 +1,6 @@
 import * as Sentry from '@sentry/react'
 import { isAdmin } from '../utils/isAdmin'
+import { getIOContext } from './ctx'
 
 if (isAdmin()) {
   Sentry.init({
@@ -21,5 +22,39 @@ if (isAdmin()) {
     // plus for 50% of sessions with an error
     replaysSessionSampleRate: 0,
     replaysOnErrorSampleRate: 0.5,
+
+    beforeSend: (event) => {
+      const ctx = getIOContext()
+
+      // Must check with false, otherwise default null's
+      // value leads to data mistakenly not sent to Sentry,
+      // which can occur if somehow we can't infer whether
+      // the apps are running under a production or development
+      // environment.
+      if (ctx.admin_production === false) {
+        const params = new URL(document?.location?.toString())?.searchParams
+        const shouldLog = params.get('forceLogs')
+
+        if (shouldLog === 'true') {
+          return makeEventWithCtx(event, ctx)
+        }
+
+        return null
+      }
+
+      return makeEventWithCtx(event, ctx)
+    },
   })
+}
+
+function makeEventWithCtx(event: any, ctx: any) {
+  const eventWithCtx = {
+    ...event,
+    tags: {
+      ...event?.tags,
+      ...ctx,
+    },
+  }
+
+  return eventWithCtx
 }